+++ docs/release_notes.rst
@@ -15,3 +15,4 @@ Features:
 * Pipelines selection using ray-tune, Optuna
 * Optimize for metrics as 'validity_index', 'davies_bouldin', 'silhouette', 'calinski_harabasz'
 * Pipelines composition from preprocessing, dimensionality reduction and clustering algorithms
+* Enable custom preprocessing, dimensionality reduction and clustering algorithms pipelines